refactor into multi files; bind config_uris to flag
